2025 day7

+++ year2025/cpp/day7/aoc.h
@@ -7,4 +7,4 @@ using std::vector;
 #pragma once
 
 int part1(vector<string> lines);
-int part2(vector<string> lines);
+long long part2(vector<string> lines);
