+++ year2025/cpp/day7/main.cpp
@@ -6,7 +6,7 @@ int main()
     const vector<string> lines = read_lines("../../../../data/day7_data.txt");
 
     int result_part1 = part1(lines);
-    int result_part2 = part2(lines);
+    long long result_part2 = part2(lines);
 
     cout << "Part 1: " << result_part1 << "\n";
     cout << "Part 2: " << result_part2 << "\n";
